int -> float 수정

+++ domain/geofenceControlInfoResponse.go
@@ -5,7 +5,7 @@ type GeofenceControlInfoResponse struct {
 	UpdateValue     int     `json:"updateValue"`
 	GeofenceGroupId int     `json:"geofenceGroupId"`
 	GeoEventType    int16   `json:"geoEventType"`
-	GeofenceRange   int     `json:"geofenceRange"`
+	GeofenceRange   float64 `json:"geofenceRange"`
 	Latitude        float64 `json:"latitude"`
 	Longitude       float64 `json:"longitude"`
 	OnTime          string  `json:"onTime"`  // formatted as yyyyMMddHHmmss
